Added configure option "--enable-extra-warnings" adding:
-W -Wno-sign-compare -Wno-unused-parameter -Wundef -Wcast-qual
-Wcast-align -Waggregate-return -Wmissing-declarations -Winline


Fixed tons of warnings, most of them were related to qualifiers,
some were aggregate returns, some other...

+++ gfxdrivers/nvidia/nvidia_overlay.c
@@ -35,7 +35,7 @@ typedef struct {
      short                 contrast;
      short                 saturation;
      short                 hue;
-     __u8                  buffer;
+     int                   buffer;
      __u32                 fbstart;
 
      struct {
